Seperated setup_data function and update test
Test is not fully updated yet
