+++ src/methods/sysid_functions/clean_sysid_output.py
@@ -64,40 +64,7 @@ def remove_highly_uncertain_points(sysid_output: Dict[str, Any], sysid_params: D
     frequency_coefficient_variation = sysid_output['Fn_poles_std'].copy()**0.5/frequencies
     damping_coefficient_variation = sysid_output['Xi_poles_std'].copy()**0.5/damping_ratios
     indices_frequency = frequency_coefficient_variation > freq_variance_treshold
-    # print(frequencies[:,-1])
-    # print('\nstd_freq')
-    # print(std_freq[:,-1])
-    # print('\nfn_poles_std')
-    # print(sysid_output['Fn_poles_std'].copy()[:,-1])
-    # print('\nvar_coef')
-    # print(frequency_coefficient_variation[:,-1])
-    # print('\n id')
-    # print(indices_frequency[:,-1])
-    # print('\nchosen')
-    # print(frequency_coefficient_variation[indices_frequency])
-    # print('\nchosen f')
-    # print(frequencies[indices_frequency])
-    # print('\ncut off')
-    # print(freq_variance_treshold)
     indices_damping   = damping_coefficient_variation > damp_variance_treshold
-    # print('\nfreq')
-    # print(frequencies[:,-1])
-    # print('\ndamp')
-    # print(damping_ratios[:,-1])
-    # print('\nstd_damp')
-    # print(std_freq[:,-1])
-    # print('\nxi_poles_std')
-    # print(sysid_output['Xi_poles_std'].copy()[:,-1])
-    # print('\nvar_coef')
-    # print(damping_coefficient_variation[:,-1])
-    # print('\n chosen')
-    # print(damping_coefficient_variation[indices_damping])
-    # print('\n id chosen')
-    # print(indices_damping[:,-1])
-    # print('\n')
-    # print(damping_ratios[indices_damping])
-    # print('\n')
-    # print(damp_variance_treshold)
     above_nyquist = frequencies > sysid_params['Fs']/2
     combined_indices = np.logical_or(np.logical_or(indices_frequency,indices_damping),above_nyquist)
     frequencies[combined_indices] = np.nan
