+++ tests/unit/methods/test_sys_id_unit.py
@@ -5,10 +5,10 @@ from datetime import datetime
 import json
 from methods.sysid import (
     sysid,
-    get_sysid_output,
-    publish_sysid_output,
-    setup_client,
+    wait_for_sysid_output,
+    publish_sysid_output
 )
+from methods.setup_data import setup_client, get_data
 from paho.mqtt.client import Client as MQTTClient
 
 
@@ -30,7 +30,7 @@ def sysid_params():
 def test_sysid_returns_expected_keys(sample_data, sysid_params):
     result = sysid(sample_data, sysid_params)
     assert isinstance(result, dict)
-    expected_keys = {'Fn_poles', 'Fn_poles_cov', 'Xi_poles', 'Xi_poles_cov', 'Phi_poles'}
+    expected_keys = {'Fn_poles', 'Fn_poles_std', 'Xi_poles', 'Xi_poles_std', 'Phi_poles','Phi_poles_std','Ufx_list'}
     assert expected_keys.issubset(result.keys())
 
 
@@ -40,36 +40,36 @@ def test_sysid_transposes_data_if_needed(sysid_params):
     assert isinstance(result, dict)
     assert "Fn_poles" in result
 
-def test_get_oma_results_success(mocker):
+def test_get_oma_results_success():
     fs = 100
     samples_needed = 600
     mock_data = np.random.randn(samples_needed, 3)
     mock_aligner = MagicMock()
     mock_aligner.extract.return_value = (mock_data, datetime.now())
 
-    result, ts = get_sysid_output(600, mock_aligner)
+    result, ts = get_data(600, mock_aligner)
 
     assert result is not None
     assert "Fn_poles" in result
 
-def test_get_oma_results_no_data(mocker):
+def test_get_oma_results_no_data():
     fs = 100
     mock_aligner = MagicMock()
     mock_aligner.extract.return_value = (np.empty((0, 3)), datetime.now())
 
-    result, ts = get_sysid_output(6000, mock_aligner)
+    result, ts = get_data(6000, mock_aligner)
 
     assert result is None
     assert ts is None
 
 
-def test_get_oma_results_not_enough_samples(mocker):
+def test_get_oma_results_not_enough_samples():
     fs = 100
     mock_aligner = MagicMock()
     data = np.random.randn(100, 3)
     mock_aligner.extract.return_value = (data, datetime.now())
 
-    result, ts = get_sysid_output(1000, mock_aligner)  # ask for too many samples
+    result, ts = get_data(1000, mock_aligner)  # ask for too many samples
 
     assert result is None
     assert ts is None
@@ -82,7 +82,7 @@ def test_get_oma_results_sysid_failure(mocker):
 
     mocker.patch("methods.sysid.sysid", side_effect=Exception("fail"))
 
-    result, ts = get_sysid_output(1, mock_aligner)
+    result, ts = get_data(1, mock_aligner)
 
     assert result is None
     assert ts is None
@@ -92,16 +92,16 @@ def test_publish_oma_results_retries_and_publishes_once(mocker):
     fs = 100
     dummy_result = {
         'Fn_poles': np.array([[3.2, 3.3], [3.4, 3.5]]),
-        'Fn_poles_cov': np.array([[0.1, 0.1], [0.1, 0.1]]),
+        'Fn_poles_std': np.array([[0.1, 0.1], [0.1, 0.1]]),
         'Xi_poles': np.array([[0.02, 0.03], [0.04, 0.05]]),
-        'Xi_poles_cov': np.array([[0.001, 0.001], [0.001, 0.001]]),
+        'Xi_poles_std': np.array([[0.001, 0.001], [0.001, 0.001]]),
         'Phi_poles': np.array([[1.0, 0.0], [0.0, 1.0]]),
     }
 
     mocker.patch("methods.sysid.time.sleep", return_value=None)
 
     mocker.patch(
-        "methods.sysid.get_sysid_output",
+        "methods.setup.get_data",
         side_effect=[
             (None, None),
             (dummy_result, datetime(2024, 1, 1))
@@ -130,16 +130,18 @@ def test_setup_client_with_multiple_topics(mocker):
     mqtt_config = {
         "host": "localhost",
         "port": 1883,
+        "ClientID": "default",
+        "userId": "",
+        "password": "",
+        "QoS":1,
         "TopicsToSubscribe": ["topic1", "topic2"]
     }
-
-    extract_mock = mocker.patch("methods.sysid.extract_fs_from_metadata", return_value=123.0)
+    extract_mock = mocker.patch("data.accel.metadata.extract_fs_from_metadata", return_value=123.0)
 
     mock_mqtt_client = MagicMock()
-    mocker.patch("methods.sysid.setup_mqtt_client", return_value=(mock_mqtt_client))
+    mocker.patch("methods.setup.setup_client", return_value=(mock_mqtt_client))
 
     client, fs = setup_client(mqtt_config)
-    client.loop_stop()
 
     extract_mock.assert_called_once_with(mqtt_config)
     client.connect.assert_called_once_with("localhost", 1883, 60)
